Update SQL query to cast count(*) (bigint, which would be interpreted as string) to int. Implement functions in /models/student/queries.ts

+++ server/models/Student/student.sql
@@ -622,7 +622,7 @@ WITH favorites_partition AS (
         upchieve.student_favorite_volunteers
 )
 SELECT
-    count(*) AS duplicates_count
+    count(*)::int AS duplicates_count
 FROM
     favorites_partition
 WHERE
